+++ app/[locale]/new/page.tsx
