+++ app/[locale]/p/[party]/page.tsx
