+++ app/[locale]/saved/SavedPageClient.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import React, { useEffect, useState } from 'react'
-import { useTranslation } from 'next-i18next'
-import { useRouter, useSearchParams } from 'next/navigation'
+import { useTranslations } from 'next-intl'
+import { useRouter, useSearchParams } from '~/i18n/navigation'
 
 // Components
 import FilterBar from '~/components/filters/FilterBar'
@@ -43,7 +43,7 @@ const SavedPageClient: React.FC<Props> = ({
   initialRecency,
   error = false
 }) => {
-  const { t } = useTranslation('common')
+  const t = useTranslations('common')
   const router = useRouter()
   const searchParams = useSearchParams()
