+++ app/[locale]/server-error/page.tsx
