+++ app/[locale]/teams/page.tsx
